Return NumOpResult when calculating fee
Currently we call the fee calculation functions `checked_div_by_foo` and
return an `Option` (to mimic std). However out `Amount` and `FeeRate`
are very different to std numeric types. We have the `NumOpResult`
already, lets use it.

While we are at it remove the `checked_` prefix since the return value
is now not `Option`.

Includes a bunch of changes to the `psbt` tests because extracting the
transaction from a PSBT has a bunch of overflow paths that need testing
caused by fee calculation.

+++ units/src/fee_rate/mod.rs
@@ -60,7 +60,7 @@ impl FeeRate {
     /// returning `None` if overflow occurred.
     pub const fn from_sat_per_kwu(sat_kwu: u64) -> Option<Self> {
         // No `map()` in const context.
-        match sat_kwu.checked_mul(4_000) {
+        match (sat_kwu as u64).checked_mul(4_000) {
             Some(fee_rate) => Some(FeeRate::from_sat_per_mvb(fee_rate)),
             None => None,
         }
